added freelance jobs api (#159)
* added freelance jobs api

* Update EVEStandard/API/FreelanceJobs.cs


---------

+++ EVEStandard/Models/API/AuthDTO.cs
@@ -5,7 +5,7 @@ namespace EVEStandard.Models.API
     public class AuthDTO
     {
         public AccessTokenDetails AccessToken { get; set; }
-        public int CharacterId { get; set; }
+        public long CharacterId { get; set; }
         public string Scopes { get; set; }
     }
 }
